Parallel milestones: C ABI + live MIDI (2.0-ready), and the stereo image fixed by mechanism
INTEGRATION (DISTRIBUTION M0+M1 + live input): new ffi/ crate -- C ABI
(pe_* / ABI v2, pe_status + pe_last_error carrying Preset::validate's
field naming across the boundary, catch_unwind everywhere, panic=abort
in a dist profile so release numbers keep their unwinding model),
boundary resampler bypassed byte-identically at 48 kHz (-104 dB alias
floors at 44.1/96, zero audio-path allocations over 800 renders), C
harness rendering sample-exactly what the CLI renders. Live MIDI:
piano-emulator --midi-in via CoreMIDI protocol ports -- the OS
translates every 1.0 source to UMP, so one parser; velocity widened to
u16 with two value-preserving lanes (old points exact, proven through
the real MIDI server via in-test virtual sources); CC64 slew; MIDI 2.0
16-bit velocity flows end to end for the SL88 MK2, with the manual
smoke steps in README.

STEREO: the inverted interchannel structure (bass -0.58 where the
recording reads +0.95, treble pan-locked) is fixed instrument-side: a
virtual capsule pair in mid/side form (mono fold-down IS the old pan --
bound at -131 dBFS on the demo), frequency-dependent board-field
coherence with its constant fixed by the Cook diffuse-field formula,
and a measured sixth-octave directivity profile of the plate's three
regimes closed by a staggered 8th-order side-channel edge matched to
the recording's 40 dB/oct transition. Stereo columns land in bench/
compass with bars built from the reference alone; gate 6-of-6 red ->
green, verified independently on fresh renders; fixed geometry is
physical (0.5 m pair -- the measured lags refuted the 12 cm guess).
Verification found and root-caused a measurement-window defect
(mid-note edge) -- fixed with one canonical window, no bar widened.

Suite 696 green / 1 red (the documented duplex alone). DECISIONS
346-391 (both reserved ranges). One declared accident: a cargo fmt
--all, reverted by proof file-by-file.

+++ engine/src/calibrate.rs
@@ -60,7 +60,7 @@ const BURST_WINDOW_S: f32 = 0.25;
 const BURST_DRAWS: u64 = 8;
 
 /// Velocity of the strike every mechanism level is quoted against.
-const REFERENCE_VELOCITY: u8 = 90;
+const REFERENCE_VELOCITY: u16 = 90;
 
 /// The key a *pedal* event's level is quoted against. The pedal recordings
 /// belong to no key, and the tuner measures them against the sampled key
