+++ tuner/src/tools/brilliance.rs
@@ -270,7 +270,7 @@ fn matched_powers(spectrum: &[f64], reference: &[f64], bands: &[(f64, f64)]) -> 
 fn render_engine(preset: &Preset, key: u8) -> Audio {
     let events = [RenderEvent::new(
         PREROLL_S as f32,
-        Event::NoteOn { key, vel: VELOCITY },
+        Event::NoteOn { key, vel: u16::from(VELOCITY) },
     )];
     let (left, right) = render_to_buffer(preset, &events, (PREROLL_S + RENDER_S) as f32);
     let skip = (PREROLL_S * f64::from(SAMPLE_RATE)) as usize;
